extract the tool call item into self contained vue component to adhere to best practices; also narrowed down the collapsed list to two items (and hid input/output while collapsed)
